examples/api/requirements.txt
# spacy-ann-linker
fastapi
uvicorn
python-dotenv